requirements.txt
# AttentionX Python Dependencies
# Install: pip install -r requirements.txt

# ── AI / ML ──────────────────────────────────────────────────────────────────
openai-whisper>=20231117          # local Whisper transcription
google-generativeai>=0.5.0        # Gemini 1.5 Flash

# ── Video & Audio processing ──────────────────────────────────────────────────
moviepy>=1.0.3                    # video trimming, cropping, caption overlay
librosa>=0.10.0                   # audio energy / loudness analysis
soundfile>=0.12.0                 # audio I/O for librosa
opencv-python-headless>=4.8.0    # frame extraction (no GUI needed)
mediapipe>=0.10.0                 # face detection & tracking

# ── Backend API ───────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9           # required for UploadFile in FastAPI

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ── Optional: GPU acceleration for Whisper (uncomment if you have NVIDIA GPU) ─
# torch>=2.0.0
# torchaudio>=2.0.0
